Merge ``origin/main`` (commits f497abf..9c03710 — phase 2 habits/practices
remediation + email-dedup) and resolve the conflicts in five routers:

- routers/goal_completions.py: keep main's idempotency check + streak diff,
  add the BUG-INFRA-024 structured-log line.
- routers/habits.py: keep main's per-habit streak population
  (``_populate_streak``) and chain it into the paginated list.
- routers/practice_sessions.py: keep main's newest-first ordering on the
  paginated list query.
- routers/practices.py: keep main's ``col(...).is_(True)`` filter.
- routers/user_practices.py: auto-merged with main's active-stage check.

Migration-drift fixes (uncovered by the new BUG-INFRA-023 alembic-check
job, now actually exercised against Postgres):

1. ``d4e5f6a7b8c9_goal_completion_unique_per_day_and_index``: the original
   index expression ``timestamp::date`` is STABLE on a TIMESTAMP WITH TIME
   ZONE column and Postgres refuses to index it.  Switched to
   ``(timestamp AT TIME ZONE 'UTC')::date`` which is IMMUTABLE and matches
   the application's UTC-everywhere invariant.  This bug was latent on
   main because no prior CI exercised the migration against Postgres.

2. ``a8b9c0d1e2f3_align_practice_duration_and_promptresponse_unique`` (new):
   resolves two real model/schema mismatches that ``alembic check``
   surfaced:
   - ``practice.default_duration_minutes`` was declared ``float`` on the
     model but ``INTEGER`` in the initial schema.  Promoted to
     DOUBLE PRECISION; existing values cast cleanly.
   - ``PromptResponse.__table_args__`` declares
     ``UniqueConstraint("user_id", "week_number")`` (BUG-JOURNAL-003) but
     no migration created it.  Added with a defensive dedup pass first.

3. ``migrations/env.py``: added an ``include_object`` filter that excludes
   four functional / partial / expression indexes whose definitions can't
   be expressed in a SQLite-compatible model declaration.  These are
   created via raw ``op.execute`` in their own migrations and remain the
   source of truth there:
   - ix_user_lower_email_unique (lower(email))
   - ix_goal_completion_unique_per_day ((timestamp AT TIME ZONE 'UTC')::date)
   - ix_goal_completion_goal_user_timestamp
   - ix_user_practice_active_stage (partial WHERE end_date IS NULL)

Verified locally with a throwaway Postgres 16 (pg_virtualenv): upgrade
head → downgrade -1 → upgrade head + alembic check all clean.  Full
backend test suite passes.

+++ backend/src/models/practice.py
@@ -9,6 +9,6 @@ class Practice(SQLModel, table=True):
     name: str = Field(max_length=255)
     description: str = Field(max_length=2_000)
     instructions: str = Field(max_length=10_000)
-    default_duration_minutes: int
+    default_duration_minutes: float
     submitted_by_user_id: int | None = Field(default=None, foreign_key="user.id")
     approved: bool = True
